Move non-platform-specific parts of IpUtils to PacketUtils and move the rest under namespace internal
