+++ Packet++/header/VlanLayer.h
@@ -3,9 +3,6 @@
 
 #include "Layer.h"
 #include "EthLayer.h"
-#if defined(WIN32) || defined(WINx64)
-#include <winsock2.h>
-#endif
 
 /// @file
