+++ Packet++/src/IPReassembly.cpp
@@ -3,7 +3,7 @@
 #include "IPReassembly.h"
 #include "IPv4Layer.h"
 #include "IPv6Layer.h"
-#include "IpUtils.h"
+#include "PacketUtils.h"
 #include "Logger.h"
 #include <string.h>
 #include "EndianPortable.h"
